pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.43.1",
    "twilio>=9.5.0",
]
